Move playback.hpp
Move playback.hpp to include/k4arecord
