+++ src/record/sdk/CMakeLists.txt
@@ -122,6 +122,7 @@ install(
     FILES
         ${K4A_INCLUDE_DIR}/k4arecord/record.h
         ${K4A_INCLUDE_DIR}/k4arecord/playback.h
+        ${K4A_INCLUDE_DIR}/k4arecord/playback.hpp
         ${K4A_INCLUDE_DIR}/k4arecord/types.h
         ${CMAKE_CURRENT_BINARY_DIR}/include/k4arecord/k4arecord_export.h
     DESTINATION
